requirements.txt
kafka-python==2.0.2
docker==6.0.1
requests==2.28.1
numpy==1.26.4
pandas==2.2.2
